[Touch FW] Fixed SAMD1x API.h File MISRA Violation - MH3-82012

+++ src/qtm_acq_samd1x_0x0009_api.h
@@ -71,7 +71,7 @@ SUBSTITUTE  GOODS,  TECHNOLOGY,  SERVICES,  OR  ANY  CLAIMS  BY  THIRD   PARTIES
 #define X(n) ((uint16_t)(1u << (n)))
 
 /* Y line bit position */
-#define Y(n) ((uint32_t)(((uint32_t)1u) << (n)))
+#define Y(n) ((uint16_t)(((uint16_t)1u) << (n)))
   
 
 /* Extract Analog / Digital Gain */
